change styles in homepage
