+++ src/App.js
@@ -1,16 +1,18 @@
 import Header from "./landingPage/Header";
-import Card from "./landingPage/Card";
+import Card from "./landingPage/Card/Card";
 import { Routes, Route } from "react-router-dom";
 import Homepage from "./Pages/Homepage";
 import MexicoPage from "./Pages/MexicoPage";
 import CroatiaPage from "./Pages/CroatiaPage";
 import SerbiaPage from "./Pages/SerbiaPage";
 import NotFoundPage from "./Pages/NotFoundPage";
+import Banner from "./landingPage/Banner/Banner";
 
 function App() {
   return (
     <div>
       <Header></Header>
+      <Banner></Banner>
       <Routes>
         <Route path="/" element={<Card />} />
         <Route path="/croatia" element={<CroatiaPage />} />
