+++ src/landingPage/Banner/Banner.css
@@ -1,0 +1,13 @@
+.container {
+  top: 50%;
+  left: 50%;
+  text-align: center;
+}
+
+h2 {
+  font-size: 40px;
+}
+
+p {
+  font-size: 30px;
+}
