+++ src/landingPage/Banner/Banner.js
@@ -1,0 +1,12 @@
+import "./Banner.css";
+
+function Banner() {
+  return (
+    <div className="container">
+      <h2>CHANGE OF SCENNERY CAN BE A GOOD THING</h2>
+      <p>Visit one of our remote offices</p>
+    </div>
+  );
+}
+
+export default Banner
