--- src/landingPage/Card.css
@@ -1,80 +1,0 @@
-.cards {
-    display: flex;
-    justify-content: space-evenly;
-    margin-bottom: 2%;
-}
-
-.card {
-    background-color: black;
-    color: var(--secondary-color);
-    border-radius: 5%;
-    width: 30%;
-    height: 400px;
-    text-align: center;
-}
-
-.card-1,
-.card-2,
-.card-3 {
-    position: relative;
-    background-size: cover;
-}
-
-.card::before {
-    content: " ";
-    opacity: 0.6;
-    display: block;
-    position: absolute;
-    left: 0;
-    top: 0;
-    width: 100%;
-    height: 100%;
-    background-size: cover;
-    border-radius: 5%;
-}
-
-.card-1::before {
-    background-image: url("https://images.unsplash.com/photo-1613601740367-410ae03b2ec7?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1170&q=80");
-}
-
-.card-2::before {
-    background-image: url("https://images.unsplash.com/photo-1599624893198-d000bdd8bef8?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1170&q=80");
-}
-
-.card-3::before {
-    background-image: url("https://images.unsplash.com/photo-1582456032629-4a54a1a65697?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1832&q=80");
-}
-
-article {
-    position: relative;
-}
-
-article>h3 {
-    font-size: 40px;
-    color: white;
-    text-align: center;
-    padding: 25% 0;
-}
-
-article>p {
-    font-size: 30px;
-    color: white;
-}
-
-.hidden {
-    display: none;
-}
-
-.card:hover h3 {
-    display: none;
-}
-
-.card:hover .hidden {
-    display: block;
-    font-weight: bold;
-    text-align: left;
-    line-height: 30px;
-    width: 90%;
-    padding: 25% 0;
-    margin-left: 5%;
-}
