+++ src/landingPage/Card/Card.css
@@ -1,0 +1,81 @@
+.cards {
+  display: flex;
+  justify-content: space-evenly;
+  margin-bottom: 2%;
+}
+
+.card {
+  background-color: black;
+  border-radius: 5%;
+  width: 30%;
+  height: 400px;
+  text-align: center;
+  position: relative;
+  color: white;
+}
+
+.card p {
+  font-size: 100%;
+  color: white;
+  background-color: rgba(0, 0, 0, 0.3);
+}
+
+.card h3 {
+  position: relative;
+  color: white;
+  top: 50%;
+  transform: translateY(-50%);
+
+}
+
+.card-1,
+.card-2,
+.card-3 {
+  position: relative;
+  background-size: cover;
+}
+
+.card::before {
+  content: " ";
+  opacity: 0.6;
+  display: block;
+  position: absolute;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  background-size: cover;
+  border-radius: 5%;
+}
+
+.card-1::before {
+  background-image: url("https://images.unsplash.com/photo-1613601740367-410ae03b2ec7?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1170&q=80");
+}
+
+.card-2::before {
+  background-image: url("https://images.unsplash.com/photo-1599624893198-d000bdd8bef8?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1170&q=80");
+}
+
+.card-3::before {
+  background-image: url("https://images.unsplash.com/photo-1582456032629-4a54a1a65697?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1832&q=80");
+}
+
+article {
+  position: relative;
+}
+
+.hidden {
+  opacity: 0;
+  text-align: left;
+  line-height: 30px;
+  bottom: 0;
+  position: absolute;
+  padding: 10px;
+  margin: 0;
+  transition: opacity 0.2s ease-in-out;
+  text-align: center;
+}
+
+.card:hover .hidden {
+  opacity: 1;
+}
